vmware: Add option to allow disabling guest quiescing
Fixes #1057: Option to disable quiesced snapshot

Added the optional setting quiesce, by default guest quiescing on
snapshot is enabled. When adding the plugin option quiesce=no the
guest quiescing on snapshot will be disabled. The plugin will
issue an error message in this case, to indicate that the backup
may be inconsistent, so the job termination will be
Backup OK -- with warnings